I modernized the cbf2-core, CE and EE build process according to docker best practices.  The image size was cut in half. The cbf2-core used to weigh in at about 780MB, now just about 500MB. The server + core used to be 4.7GB, now it is 2.56GB. The build is also twice as fast.
Containers built on any given host machine with docker installed
will assume the UID and GID of the current, non-privileged user on the system.
The pentaho user will still be used, but will share this UID:GID.
This has the advantage that any files created by the container will be accessible
using the username on the host computer. So this should be an improvement wtih no downside.

I also re-wrote the handling of EE images to handle the case where
I extract and install the baserver + plugins, then patch and zip the thing up.
I have not fixed the 'apply EE patches' part of the script.  Patching is now
cummulative, so that should be much easier to update at some point.